requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
decorator==5.0.9
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
flake8==3.9.2
idna==3.2
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
